backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.0
openpyxl==3.1.2
openai==1.12.0
python-dotenv==1.0.0